fixed paths to dataset
if we included the datasets in our repository, github would no longer allow us to push the files. So we must each have the datasets stored outside the repository. To reference them, the datasets should be located in their original folder from downloading, then all stored in a "datasets" folder, in the same folder as the "grand-grand-parent" of the project code